rename: 공통 SVG 컴포넌트 폴더 이동
