+++ components/common/icons/Avatar.tsx
@@ -1,10 +1,7 @@
-type Props = { width: number; height: number };
-
-export default function ImageIcon({ width, height }: Props) {
+export default function Avatar(props: React.SVGProps<SVGSVGElement>) {
   return (
     <svg
-      width={width}
-      height={height}
+      {...props}
       viewBox='0 0 84 84'
       fill='none'
       xmlns='http://www.w3.org/2000/svg'
