+++ components/common/icons/SlideNextIcon.tsx
@@ -1,8 +1,7 @@
-export default function NextIcon() {
+export default function SlideNextIcon(props: React.SVGProps<SVGSVGElement>) {
   return (
     <svg
-      width='24'
-      height='24'
+      {...props}
       viewBox='0 0 24 24'
       fill='none'
       xmlns='http://www.w3.org/2000/svg'
